PHP 8.2 | BCFile::getMethodParameters(): sync with upstream / add support for true pseudotype
As pulled in upstream PR squizlabs/PHP_CodeSniffer3662 / PHPCSStandards/PHP_CodeSniffer 49.

Includes unit tests.
Includes moving some tests to allow the tests to still work as expected.

Support for the `true` type was previously already added to the `FunctionDeclarations::getParameters()` method in PR 368.

Refs:
* squizlabs/PHP_CodeSniffer 3662
* PHPCSStandards/PHP_CodeSniffer 49
* PHPCSStandards/PHPCSUtils 368
